Add confirmation dialog before deleting a server

+++ app/src/main/java/tv/olaris/android/ui/serverList/ServerItemAdapter.kt
@@ -9,6 +9,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import kotlinx.coroutines.launch
 import tv.olaris.android.OlarisApplication
 import tv.olaris.android.R
@@ -18,6 +19,7 @@ class ServerItemAdapter(context: Context) : ListAdapter<Server, ServerItemAdapte
     class ServerItemHolder(val view: View) : RecyclerView.ViewHolder(view){
         val serverUrl = view.findViewById<TextView>(R.id.text_server_label)
         val deleteServerIcon = view.findViewById<ImageView>(R.id.icon_delete_server)
+        val context = view.context
 
     }
 
@@ -29,9 +31,21 @@ class ServerItemAdapter(context: Context) : ListAdapter<Server, ServerItemAdapte
     override fun onBindViewHolder(holder: ServerItemHolder, position: Int) {
         holder.serverUrl.text = getItem(position).name
         holder.deleteServerIcon.setOnClickListener{
-            OlarisApplication.applicationContext().applicationScope.launch {
-                OlarisApplication.applicationContext().serversRepository.deleteServer(getItem(position))
-            }
+
+            MaterialAlertDialogBuilder(holder.context)
+                    .setTitle("Remove server from application")
+                    .setMessage("Are you sure you want to delete the server '${holder.serverUrl.text}' from the app?")
+                    .setNeutralButton("Cancel") { dialog, which ->
+                    }
+
+                    .setPositiveButton("Confirm") { dialog, which ->
+                        OlarisApplication.applicationContext().applicationScope.launch {
+                            OlarisApplication.applicationContext().serversRepository.deleteServer(getItem(position))
+                        }
+                    }
+                    .show()
+
+
         }
     }
 }
